requirements.txt
pandas
numpy
matplotlib
requests
python-dotenv
plotly
pyyaml
beautifulsoup4
